Move `tests` module into `lib.rs` to fix broken unused import warning
